backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
langchain==0.3.1
langchain-community==0.3.1
langchain-groq==0.2.0
langchain-google-genai==2.0.0
chromadb==0.5.11
sentence-transformers==3.1.1
pypdf==5.0.1
pillow==10.4.0
python-multipart==0.0.9